integrate into colorGrowth.py and delete class file
