+++ scripts/imgAndVideo/colorGrowth.py
@@ -28,7 +28,6 @@
 import datetime, random, argparse, ast, os.path
 import numpy as np
 from PIL import Image
-import colorpoop as cp		# Uses my custom class in colorpoop.py
 # import sys
 
 # ARGUMENT PARSING AND GLOBALS.
@@ -72,6 +71,40 @@ if failedMutationsThreshold == None:
 terminatePaintingAtFillCount = int(allesPixelCount * stopPaintingPercent)
 
 print('Will generate ', numIMGsToMake, ' image(s).')
+
+class coordinate:	# for "color poop," formerly "Coordinate:"
+	# slots for allegedly higher efficiency re: https://stackoverflow.com/a/49789270
+	__slots__ = ["XYtuple", "maxX", "maxY", "RGBcolor", "isAlive", "isConsumed", "emptyNeighbors"]
+	def __init__(self, x, y, maxX, maxY, RGBcolor, isAlive, isConsumed, emptyNeighbors):
+		self.XYtuple = (x, y)
+		self.RGBcolor = RGBcolor; self.isAlive = isAlive;	self.isConsumed = isConsumed
+		# Adding all possible empty neighbor values even if they would result in values out of bounds of image (negative or past maxX or maxY), and will check for and clean up pairs with out of bounds values after:
+		tmpList = [ (x-1, y-1), (x-1, y), (x-1, y+1), (x, y-1), (x, y+1), (x+1, y-1), (x+1, y), (x+1, y+1) ]
+		deleteList = []
+		for element in tmpList:
+			if -1 in element:
+				deleteList.append(element)
+		for element in tmpList:		# TO DO: debug whether I even need this; the print never happens:
+			if (maxX+1) in element:
+				deleteList.append(element)
+		for element in tmpList:
+			if (maxY+1) in element:
+				deleteList.append(element)
+		# reduce deleteList to a list of unique tuples (in case of duplicates, which can lead us to attempt to remove something that ins't there, which throws an exception and stops the script) :
+		deleteList = list(set(deleteList))
+		# the deletions:
+		for no in deleteList:
+			tmpList.remove(no)
+		# finallu initialize the intended object member from that built list:
+		self.emptyNeighbors = list(tmpList)
+	def getRNDemptyNeighbors(self):
+		random.shuffle(self.emptyNeighbors)		# shuffle the list of empty neighbor coordinates
+		nNeighborsToReturn = np.random.random_integers(0, len(self.emptyNeighbors))		# Decide how many to pick
+		rndNeighborsToReturn = []		# init an empty array we'll populate with neighbors and return
+		# iterate over nNeighborsToReturn items in shuffled self.emptyNeighbors and add them to a list to return:
+		for pick in range(0, nNeighborsToReturn):
+			rndNeighborsToReturn.append(self.emptyNeighbors[pick])
+		return rndNeighborsToReturn
 
 	# function takes two ints and shifts each up or down one or not at all. I know, it doesn't receive a tuple as input but it gives one as output:
 # TO DO: use the following repeatedly only if Coordinate.getRNDemptyNeighbors() fails:
@@ -125,7 +158,7 @@ for n in range(1, (numIMGsToMake + 1) ):		# + 1 because it iterates n *after* th
 	arr = []	# list of Coordinate objects
 	for xCoord in range(0, width):
 		for yCoord in range(0, height):	# RGBcolor can also be initialized with: np.random.randint(0, 255, size=3)
-			arr.append(cp.coordinate(xCoord, yCoord, width, height, backgroundColor, False, False, None))
+			arr.append(coordinate(xCoord, yCoord, width, height, backgroundColor, False, False, None))
 
 	unusedCoords = []		# list of tuples of unused coordinates
 	for coord in arr:
